修改代码告警



# message auto-generated for no-merge-commit merge:
!158 merge master into master

修改代码告警

Created-by: M__Cheng
Commit-by: zhangcheng
Merged-by: turing_project1
Description: ## 类型标签
<!--  [ ] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [x] 代码重构
- [ ] 文档更新
- [ ] 其他，请描述：

## 描述
<!--简要描述本次改动的背景，包括改动的原因、解决的问题等。-->
修改代码告警，禁止数字直接转成指针
## 测试项
<!--描述进行了哪些测试来验证本次改动，或新增了哪些测试用例。-->
- TestAdxlEngineTransferAsync
- TestAdxlEngineTransferAsyncWithMultiThread
## 测试结果
<!--描述上述测试项的测试结果，可通过表格、图片等形式展示。-->
程序运行结果符合预期
## Checklist
<!--  [x] 表示选中 -->
- [x] 该Pull Request的代码风格和项目的代码风格一致
- [x] 提交的代码已经过充分的验证
- [x] 涉及的相关文档修改已更新
- [x] 标题中正确使用了类型标签（例如：feat/bugfix/refactor/docs/test等）

## 其它（可选）
<!--可以在这里描述任何和该Pull Request相关的说明。-->



See merge request: cann/hixl-dev!158

+++ src/llm_datadist/adxl/adxl_inner_engine.cc
@@ -299,7 +299,7 @@ Status AdxlInnerEngine::TransferAsync(const AscendString &remote_engine,
                            "Failed to get channel, remote_engine:%s", remote_engine.GetString());
   std::lock_guard<std::mutex> transfer_lock(channel->GetTransferMutex());
   auto id = next_req_id_.fetch_add(1);
-  req = reinterpret_cast<void*>(id);
+  req = reinterpret_cast<void*>(static_cast<uintptr_t>(id));
   ADXL_CHK_STATUS_RET(channel->TransferAsync(operation, op_descs, optional_args, req),
                       "Failed to transfer async, remote_engine:%s", remote_engine.GetString());
   req2channel_[id] = remote_engine;
@@ -308,7 +308,7 @@ Status AdxlInnerEngine::TransferAsync(const AscendString &remote_engine,
 
 Status AdxlInnerEngine::GetTransferStatus(const TransferReq &req, TransferStatus &status) {
   std::lock_guard<std::mutex> lock(mutex_);
-  auto id = reinterpret_cast<uint64_t>(req);
+  auto id = static_cast<uint64_t>(reinterpret_cast<uintptr_t>(req));
   auto it = req2channel_.find(id);
   if(it == req2channel_.end()) {
     LLMLOGE(FAILED, "Request not found, request has been completed or does not exist, req: %llu", id);
